Fix for FAT, exFAT and NTFS
FAT, exFAT and NTFS don't allow double quotes (") in file names. This prevents git cloning the repo or downloading this 2 wallpapers on any drive using any of those FS. This is specially problematic for Windows users because afaik those are the only FS supported on Windows. Renaming this files to remove the (") fixes this and as a side effect also aligns them with the naming other wallpapers with multiple color variants. More information on why (") are not supported on those FS can be found here: https://en.wikipedia.org/wiki/Filename#Problematic_characters